monit-chat-api/requirements.txt
fastapi[standard]==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic[email]==2.8.2
motor>=3.7.1
python-dotenv==1.0.0
pymongo==4.14.0
openai
sqlparse==0.5.3
bcrypt
PyJWT
psycopg2-binary